Now it compiles and doesn't crash
